+++ gnrpy/RELEASE_NOTES.rst
@@ -3,8 +3,8 @@ UPCOMING RELEASE
 
 TBD
 
-Version 24.5.30
-===============
+Version 24.5.30.1
+=================
 
 
 New Features